requirements.txt
google-genai
opencv-python
python-dotenv
numpy
Pillow
streamlit